resolves #677 refactor measurement value helpers into module
- refactor measurement value helpers into Measurements module
- mixin Measurements module where needed
- raise argument exception in to_pt if units are not recognized
- allow units to be passed to to_pt as symbol
